SkillSage-backend/requirements.txt
fastapi
uvicorn
langchain
pypdf
faiss-cpu
sentence-transformers
python-dotenv
openai
PyPDF2